refactor: consolidate CLI docs and extract shell builtins dispatcher
