+++ src/cli/libmilkscript/CLIcore_UI_execute_internal.h
@@ -121,4 +121,6 @@ int cli_handle_herestring_early(errno_t *retval);
  */
 int cli_handle_background(errno_t *retval);
 
+int cli_handle_shell_builtins(void);
+
 #endif /* CLICORE_UI_EXECUTE_INTERNAL_H */
